# config/passport.js
const passport = require("passport");
const GoogleStrategy = require("passport-google-oauth20").Strategy;
const User = require("../models/userSchema");

// Serialize user into the session
passport.serializeUser((user, done) => {
  console.log("Serializing user:", user); // Log the user being serialized
  done(null, user.id); // Store user ID in the session
});

// Deserialize user from the session
passport.deserializeUser(async (id, done) => {
  try {
    const user = await User.findById(id); // Retrieve user by ID from MongoDB
    console.log("Deserialized user:", user); // Log the deserialized user
    done(null, user); // Attach user object to req.user
  } catch (error) {
    console.error("Error deserializing user:", error); // Log any errors during deserialization
    done(error, null);
  }
});

// Google Strategy
passport.use(
  new GoogleStrategy(
    {
      clientID: process.env.GOOGLE_CLIENT_ID,
      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
      callbackURL:
        process.env.NODE_ENV === "production"
          ? process.env.GOOGLE_CALLBACK_URL_PROD
          : process.env.GOOGLE_CALLBACK_URL_LOCAL,
    },
    async (accessToken, refreshToken, profile, done) => {
      console.log("Google OAuth strategy called"); 
      try {
        let user = await User.findOne({ email: profile.emails[0].value });

        if (!user) {
          user = new User({
            email: profile.emails[0].value,
            firstName: profile.name.givenName,
            lastName: profile.name.familyName,
            role: "user",
          });
          await user.save();
        }
        //console.log("User authenticated:", user);  Debug log
        return done(null, user);
      } catch (error) {
        return done(error, null);
      }
    }
  )
);